// apps/web/src/lib/api/model/query-group.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * edcurate-api
 * OpenAPI spec version: 0.1.0
 */
import type { SavedResourceResponse } from './saved-resource-response';

/**
 * A group of saved resources from the same search query.
 */
export interface QueryGroup {
  search_query: string;
  items: SavedResourceResponse[];
}